# Conflicts:
#	src/main/java/org/choongang/global/AdvancedSearch.java
#	src/main/java/org/choongang/thesis/controllers/ThesisController.java
#	src/main/java/org/choongang/thesis/controllers/ThesisSearch.java
#	src/main/java/org/choongang/thesis/entities/ThesisView.java
#	src/main/java/org/choongang/thesis/entities/ThesisViewDaily.java
#	src/main/java/org/choongang/thesis/entities/ThesisViewDailyId.java
#	src/main/java/org/choongang/thesis/entities/ThesisViewId.java
#	src/main/java/org/choongang/thesis/repositories/ThesisViewDailyRepository.java
#	src/main/java/org/choongang/thesis/repositories/ThesisViewRepository.java
#	src/main/java/org/choongang/thesis/services/ThesisInfoService.java
#	src/main/java/org/choongang/thesis/services/ThesisViewService.java
#	src/main/java/org/choongang/thesisAdvance/services/TrendInfoService.java

+++ src/main/java/org/choongang/thesis/entities/ThesisViewDailyId.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ThesisViewDailyId {
-    private int tid;
+    private Long tid;
     private int uid;
     private LocalDate date;
 }
